Move TelescopeTypes to separate file
