+++ oaebu_workflows/dags/jstor_telescope.py
@@ -17,7 +17,7 @@
 # The keywords airflow and DAG are required to load the DAGs from this file, see bullet 2 in the Apache Airflow FAQ:
 # https://airflow.apache.org/docs/stable/faq.html
 
-from oaebu_workflows.config import TelescopeTypes
+from oaebu_workflows.utils.identifiers import TelescopeTypes
 from oaebu_workflows.workflows.jstor_telescope import JstorTelescope
 from observatory.platform.utils.workflow_utils import make_observatory_api
