+++ oaebu_workflows/dags/onix_telescope.py
@@ -17,8 +17,7 @@
 # The keywords airflow and DAG are required to load the DAGs from this file, see bullet 2 in the Apache Airflow FAQ:
 # https://airflow.apache.org/docs/stable/faq.html
 
-
-from oaebu_workflows.config import TelescopeTypes
+from oaebu_workflows.utils.identifiers import TelescopeTypes
 from oaebu_workflows.workflows.onix_telescope import OnixTelescope
 from observatory.platform.utils.workflow_utils import make_observatory_api
