+++ oaebu_workflows/utils/identifiers.py
@@ -1,0 +1,28 @@
+# Copyright 2020 Curtin University
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# Author: James Diprose
+
+class TelescopeTypes:
+    """ The type_id's for known TelescopeTypes """
+
+    fulcrum = "fulcrum"
+    google_analytics = "google_analytics"
+    google_books = "google_books"
+    jstor = "jstor"
+    oapen_irus_uk = "oapen_irus_uk"
+    onix = "onix"
+    onix_workflow = "onix_workflow"
+    ucl_discovery = "ucl_discovery"
+    wos = "wos"
