+++ oaebu_workflows/workflows/tests/test_ucl_discovery_telescope.py
@@ -24,7 +24,7 @@ from airflow.exceptions import AirflowException, AirflowSkipException
 from airflow.models.connection import Connection
 from click.testing import CliRunner
 from croniter import croniter
-from oaebu_workflows.config import TelescopeTypes
+from oaebu_workflows.utils.identifiers import TelescopeTypes
 from observatory.api.client.model.organisation import Organisation
 from observatory.api.server import orm
 from oaebu_workflows.workflows.ucl_discovery_telescope import (
